Move the pattern to MaterializeEncoding pass cleanup.
