+++ compiler/src/iree/compiler/Codegen/Common/ConvertToDestinationPassingStylePass.cpp
@@ -17,8 +17,14 @@
 #include "iree/compiler/Codegen/Dialect/Codegen/IR/IREECodegenOps.h"
 #include "iree/compiler/Codegen/Transforms/Transforms.h"
 #include "iree/compiler/Codegen/Utils/Utils.h"
+#include "iree/compiler/Dialect/HAL/IR/HALOps.h"
+#include "iree/compiler/Dialect/LinalgExt/IR/LinalgExtOps.h"
+#include "iree/compiler/Dialect/Util/IR/UtilDialect.h"
+#include "iree/compiler/Dialect/Util/IR/UtilOps.h"
 #include "llvm/ADT/DenseSet.h"
+#include "llvm/ADT/EquivalenceClasses.h"
 #include "llvm/ADT/STLExtras.h"
+#include "llvm/ADT/StringSet.h"
 #include "llvm/ADT/TypeSwitch.h"
 #include "mlir/Analysis/SliceAnalysis.h"
 #include "mlir/Dialect/Arith/IR/Arith.h"
@@ -37,6 +43,7 @@
 #include "mlir/Pass/Pass.h"
 #include "mlir/Pass/PassManager.h"
 #include "mlir/Transforms/GreedyPatternRewriteDriver.h"
+#include "mlir/Transforms/Passes.h"
 
 namespace mlir::iree_compiler {
 
@@ -494,7 +501,7 @@ replaceUnpackEmptyWithAllocTensor(OpBuilder &b,
 }
 
 namespace {
-struct RemoveSplatCstOutsDependency
+struct RemoveCstOutsDependency
     : public OpInterfaceRewritePattern<linalg::LinalgOp> {
   using OpInterfaceRewritePattern<linalg::LinalgOp>::OpInterfaceRewritePattern;
 
@@ -521,86 +528,6 @@ struct RemoveSplatCstOutsDependency
       Value fillOp =
           rewriter.create<linalg::FillOp>(loc, cstOp, emptyTensor).result();
       op->setOperand(opOperand.getOperandNumber(), fillOp);
-    }
-    if (!modifiedOutput) {
-      rewriter.cancelOpModification(op);
-      return failure();
-    }
-    rewriter.finalizeOpModification(op);
-    return success();
-  }
-};
-
-/// Converts contraction ops that have readonly init to the below form. It is a
-/// special case in data-tiling fusion concept that materializes encodings into
-/// identity layout. In the context, the matmul op could use readonly in init
-/// tensors and then gets stored into the writeonly buffer. With the rewrite,
-/// the iter_arg takes tensor.empty() instead, and avoids writing data outside
-/// forall op after distribution.
-///
-/// %0 = tensor.empty()
-/// %1 = linalg.fill %1(%cst_0) outs(%0)
-/// %2 = linalg.original_op ins(%lhs, %rhs) outs(%1)
-/// %3 = linalg.elementwise_add ins(%2, %acc) outs(%0)
-struct RemoveReadOnlyOutsDependencyForContractionOps
-    : public OpInterfaceRewritePattern<linalg::LinalgOp> {
-  using OpInterfaceRewritePattern<linalg::LinalgOp>::OpInterfaceRewritePattern;
-
-  LogicalResult matchAndRewrite(linalg::LinalgOp op,
-                                PatternRewriter &rewriter) const override {
-    if (!linalg::isaContractionOpInterface(op)) {
-      return failure();
-    }
-    rewriter.startOpModification(op);
-    bool modifiedOutput = false;
-    Location loc = op.getLoc();
-    for (OpOperand &opOperand : op.getDpsInitsMutable()) {
-      auto loadOp = opOperand.get()
-                        .getDefiningOp<IREE::TensorExt::DispatchTensorLoadOp>();
-      if (!loadOp) {
-        continue;
-      }
-      auto access = llvm::cast<IREE::TensorExt::DispatchTensorType>(
-                        loadOp.getSource().getType())
-                        .getAccess();
-      if (access != IREE::TensorExt::TensorAccess::ReadOnly) {
-        continue;
-      }
-      modifiedOutput = true;
-
-      RankedTensorType type = loadOp.getResult().getType();
-      int64_t outputRank = type.getRank();
-      SmallVector<utils::IteratorType> iterators(outputRank,
-                                                 utils::IteratorType::parallel);
-      SmallVector<AffineMap> maps(3,
-                                  rewriter.getMultiDimIdentityMap(outputRank));
-      SmallVector<OpFoldResult> mixedSizes =
-          tensor::getMixedSizes(rewriter, loc, opOperand.get());
-      Type elementType = type.getElementType();
-      Value initOp =
-          rewriter.create<tensor::EmptyOp>(loc, mixedSizes, elementType);
-      Value zero = rewriter.create<arith::ConstantOp>(
-          loc, rewriter.getZeroAttr(elementType));
-      Value fill = rewriter.create<linalg::FillOp>(loc, zero, initOp).result();
-
-      // Create a generic op to add back the original output tensor operand.
-      rewriter.setInsertionPointAfter(op);
-      auto genericOp = rewriter.create<linalg::GenericOp>(
-          loc, type, ValueRange{op->getResult(0), opOperand.get()}, initOp,
-          maps, iterators,
-          [&](OpBuilder &b, Location nestedLoc, ValueRange args) {
-            Value result;
-            if (llvm::isa<FloatType>(elementType)) {
-              result = b.create<arith::AddFOp>(nestedLoc, args[0], args[1]);
-            } else {
-              result = b.create<arith::AddIOp>(nestedLoc, args[0], args[1]);
-            }
-            b.create<linalg::YieldOp>(nestedLoc, result);
-          });
-
-      // Update the contraction op to use the new zero tensor as output operand.
-      rewriter.modifyOpInPlace(op, [&]() { op.setDpsInitOperand(0, fill); });
-      op->getResult(0).replaceAllUsesExcept(genericOp->getResult(0), genericOp);
     }
     if (!modifiedOutput) {
       rewriter.cancelOpModification(op);
@@ -699,8 +626,7 @@ void ConvertToDestinationPassingStylePass::runOnOperation() {
 
   {
     RewritePatternSet patterns(context);
-    patterns.insert<RemoveSplatCstOutsDependency>(context, /*benefits=*/10);
-    patterns.insert<RemoveReadOnlyOutsDependencyForContractionOps>(context);
+    patterns.insert<RemoveCstOutsDependency>(context);
     if (failed(applyPatternsGreedily(funcOp, std::move(patterns)))) {
       return signalPassFailure();
     }
