+++ compiler/src/iree/compiler/Codegen/LLVMCPU/Passes.cpp
@@ -128,7 +128,6 @@ addTileAndDistributePasses(OpPassManager &funcPassManager,
     return;
   }
   if (clTileDispatchUsingForall) {
-    funcPassManager.addPass(createConvertToDestinationPassingStylePass());
     funcPassManager.addPass(
         createTileAndDistributeToWorkgroupsUsingForallOpPass());
     funcPassManager.addPass(createBufferizeDispatchTensorLoadStorePass());
